requirements.txt
pyspark==3.5.1
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
pandas==2.2.0
numpy>=1.24.3
tqdm==4.66.1
lxml==4.9.3
requests==2.31.0